+++ README.md
@@ -1,7 +1,7 @@
 # LANscape
 A python based local network scanner.
 
-![screenshot](https://github.com/mdennis281/py-lanscape/raw/main/src/lanscape/static/img/readme1.png)
+![screenshot](https://github.com/mdennis281/py-lanscape/raw/main/src/lanscape/ui/static/img/readme1.png)
 
 ## Local Run
 ```sh
