+++ src/lanscape/ui/static/css/style.css
