+++ src/lanscape/ui/static/img/ico/apple-touch-icon.png
