+++ src/lanscape/ui/static/img/ico/site.webmanifest
