+++ src/lanscape/ui/static/js/core.js
