+++ src/lanscape/ui/static/js/layout-sizing.js
