+++ src/lanscape/ui/static/js/quietReload.js
