+++ src/lanscape/ui/static/js/shutdown-server.js
