+++ src/lanscape/ui/static/js/subnet-info.js
