+++ src/lanscape/ui/static/lanscape.webmanifest
