+++ src/lanscape/ui/templates/core/head.html
