+++ src/lanscape/ui/templates/core/scripts.html
