+++ src/lanscape/ui/templates/info.html
