+++ src/lanscape/ui/templates/scan.html
