+++ src/lanscape/ui/templates/scan/ip-table-row.html
