+++ src/lanscape/ui/templates/scan/ip-table.html
