+++ src/lanscape/ui/templates/shutdown.html
